Re-testing highlighted issues in  WAVE toolbar
We have retested the differences pointed out in issue #3. Some of the problems
have been resolved since we've tested the WAVE toolbar, but we feel some of
the issues, even though indicated are still not categories as found.

* Large text does not have a contrast ratio of at least 3:1 so does not meet AA (you must select the Contrast tab in WAVE to see these identified)
* Large text does not have a contrast ratio of at least 4.5:1 so does not meet AAA (you must select the Contrast tab in WAVE to see these identified)
* html element is missing a lang attribute (identified at top of the page)
* Missing page title (identified at top of the page)
* Missing H1 (identified at top of the page)
* Empty label found

Even though WAVE indicates the double column headers, it doesn't raise it as an
issue and explain how it could be fixed.

We marked it as a warning but we still feel the warning message is not strong
enough and does not include a call to action, even though the problem and the
fix is clearly explained in the documentation.

Fixes #3