Improve table examples to be more realistic
Two tables are only filled with random numbers and letters,
not giving tools a lot of context.
This changes that to more realistic examples.

And one table was for a test needing the scope attribute
while the example itself would be fine without it.
This changes that to a table which has a bit more of a need for it.
(Inspired by first example on https://www.w3.org/WAI/tutorials/tables/two-headers/)