Remove font element from 'deprecated color definition' test
The 'deprecated font element' is already there to test for the font element.
This test is for checking if colour has been set within the HTML
and not within CSS.

This changes that by removing the font element and adding inline styles instead
so that this test is only about the colour and not about the font element as well.